+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning].
 
 ## [Unreleased]
 
+## [0.2.3] - 2024-05-12
+
 ### Added
 
 - Add a mount target to the base controller. ([@skryukov])
@@ -34,7 +36,8 @@ and this project adheres to [Semantic Versioning].
 
 [@skryukov]: https://github.com/skryukov
 
-[Unreleased]: https://github.com/skryukov/turbo-mount/compare/v0.2.2...HEAD
+[Unreleased]: https://github.com/skryukov/turbo-mount/compare/v0.2.3...HEAD
+[0.2.3]: https://github.com/skryukov/turbo-mount/commits/v0.2.3
 [0.2.2]: https://github.com/skryukov/turbo-mount/commits/v0.2.2
 [0.2.0]: https://github.com/skryukov/turbo-mount/commits/v0.2.0
 [0.1.0]: https://github.com/skryukov/turbo-mount/commits/v0.1.0